men with families
I looked into the question of how men with families fared. The idea is
that you can predict whether a man with family will survive by looking
at whether his family survived or not.

What we find is that if a man’s family members do not survive he does
not survive. Chivalry.